+++ src/components/business/TextAlign/index.jsx
@@ -14,12 +14,13 @@ export default class TextAlign extends React.Component {
 
   setAlignment = (e) => {
     this.props.editor.toggleSelectionAlignment(e.currentTarget.dataset.alignment)
+    this.props.editor.requestFocus()
   }
 
   render () {
 
     const { currentAlignment } = this.state
-    const { language, textAlignMaps} = this.props
+    const { language, textAlignOptions} = this.props
     const textAlignmentTitles = [
       language.controls.alignLeft,
       language.controls.alignCenter,
@@ -30,7 +31,7 @@ export default class TextAlign extends React.Component {
     return (
       <div className="control-item-group">
         {
-          textAlignMaps.map((item, index) => {
+          textAlignOptions.map((item, index) => {
             return (
               <button
                 type="button"