+++ src/languages/zh.js
@@ -33,6 +33,7 @@ export default {
     floatRight: '右浮动',
     superScript: '上标',
     subScript: '下标',
+    removeStyles: '清除样式',
     headings: '标题',
     header: '标题',
     normal: '常规',
@@ -42,6 +43,7 @@ export default {
     code: '代码',
     link: '链接',
     unlink: '清除链接',
+    hr: '水平线',
     media: '媒体',
     emoji: '小表情'
   },